collection-service/functions/add_card/requirements.txt
requests
boto3
botocore
aws_xray_sdk
python-jose